Fix :user in the issues URI
The connector builds the URI to the issues with the user's GitHub ID.
When you try to connect to the issues in a repo of another user, it
fails with a HTTP 404 error: Not Found.

So, the URI should be constructed with the user in the repo URL.  This
fix it, and now the connector can retreive issues from repositories of
other users.